finish building test suites
The test suite build & data are pretty near their final form.

It's time to move some of the code into the effio lib so this can
be two tools, one for building the test suite, another for executing.
That way, the same code can be used for both of those and the graph
generation.