Rewrite summarization to be correct for ddirs.
Discovered a few bugs that I masked earlier by returning slices
from the subsamples/histograms.

This is now correct but a little slower since it's doing ~3 passes
over the data. The 3rd pass is sorting a []float of just the values
so not quite a full pass like the other two.