Try refactoring suite commands to reuse flag parsing code.
The code was copied all over the place. This attempts to bring it
all into cmd_suite.go and make it reusable. It's not quite right
at this point due to the way Go handles mixin structs and copying
of structs when initializing.

To finish this, it looks like each subcommand will have to call a
function to add flags to the flagset at the last moment before
calling flagset.Parse(). Maybe.

Throwing this in a branch for now to rot. This whole tool is getting
restructured so it may not be worth completing this work after all.