Refactor summarization code & command.
The structure of the new code is still fairly similar to the old but
most lines have been touched due to the need for a lot more type
casting.

A few functions were broken for repeatability, namely for doing
summaries on histogram buckets that resemble the top-level summary.

Manually tested on the biggest file available. It used 12GB of RAM and
ran for 5m:48s. 1m:58s of that time was reading the CSV. Previously,
this data file would run effio into OOM on a machine with 16GB of RAM.

* since plotinum is out, there's no good reason to use float64 values
* switched LatRec from float64 values to uint32
* switched lots of other fields to more appropriate types
* renamed a few structs & variables for (hopefully) more consistency
* added LatHgram & LatBktSmry for histograms & bin summaries
* removed hacks around histograms
* updated cmd_histogram.go to follow changes in summarize_lat_log.go
* update test.go to use updated types
* moved CSV file handling code to its own file, fio_output_csv.go
* added a MarshalJSON for map[float64]*LatRec that sorts keys