Clean up record sorting & export more json.
Record sorting was causing weird results in the graphs so
now the summary is a bit safer about it and does multiple
sorts in the percentiles() function so the data is always
in time order again when it returns. This is a bit expensive
but doesn't seem to hurt runtime too badly.

Also started exporting other fields into the JSON that might
be useful in graphs.