fix nasty bug, remove second dimension, clean layout
The nasty bug was that log_type was not being inspected for the main
line graph which caused bw/iops to be inserted ahead of latency on some
devices which is totally wrong.

Removed the area graph because it looked awful and now the layout of
controls is a little nicer since all the small form elements are now
in the right-hand side column.