Check in final tweaks before Summit 2014 talk.
I forgot to commit before the talk and then didn't open my laptop
for a week.

Most of the changes are adding the Summit slide backgrounds and cleaning
up all the coordinates for impress.js.

While at it I also noticed I wasn't using the javascript promises
quite right and fixed that up.